Generate Or Enhancement Cover Letter/requirements.txt
fastapi==0.78.0
uvicorn==0.18.3
groq==0.5.0
requests==2.31.0
pandas==1.3.5
numpy==1.21.6
python-dotenv==0.21.0
pydantic==1.10.12
httpx>=0.24,<0.26
